pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi-users[sqlalchemy]>=15.0.4",
    "fastapi[standard]>=0.135.1",
    "langchain>=1.2.13",
    "langchain-openai>=1.1.11",
    "psycopg2>=2.9.11",
    "pyjwt>=2.12.1",
    "sqlmodel>=0.0.37",
]
